Updates to Xposed module and ForceDoze service
New:

* Option to disable motion sensing for Doze, without disabling device
motion sensors. This feature is EXPERIMENTAL and requires Xposed. Please
let me know if this option works as intended or not

* Changed:

* ForceDoze now ignores lockscreen timeout value by default, so a wake
lock is no longer required. If this has unwanted effects, please
disable 'Ignore lockscreen timeout' in Settings. PS: You might have to
manually turn it on if you're updating the app and this isn't the first
install