Bug fixes
* Fixed: Crash when clicking on Donate option.
* Fixed: ForceDoze not adding itself to Doze whitelist on Nougat. Now,
it will show you a notification that will take you to the battery
optimisation settings to manually add ForceDoze, as it's not possible to
add it automatically by ForceDoze on Nougat.
* Fixed: ForceDoze calling deviceidle to add itself to whitelist,
causing an exception because DEVICE_POWER permission wasn't granted. The
permission cannot be granted to apps like ForceDoze, hence a workaround
has been implemented as described above.